feat: moved getquestionnairesfilled to utils/questionnaires.js
